overview_zh.md update 修复server对client的文件操作

+++ core/option/exsync/server/scan.go
@@ -3,6 +3,6 @@ package serverOption
 type VerifyManage struct {
 	AesKey      string
 	RemoteID    string
-	Offset      int
+	Offset      int64
 	Permissions map[string]struct{}
 }
